pyproject.toml
[project]
name = "pat-auth-fastapi-demo"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.0.0",
    "bcrypt>=4.0.0,<5.0.0",
    "alembic>=1.14.0",
    "pyjwt>=2.11.0",
    "email-validator>=2.3.0",
    "flowio>=1.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "pre-commit>=4.0.0",
]
